delete record

+++ pages/api/fundraisers/[id].ts
@@ -1,25 +1,23 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import prisma from "../../../lib/prisma";
 
-// DELETE /api/fundraiser/:id
 export default async function handle(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const fundraiserId = req;
+  const fundraiserId = req.query.id?.toString();
+  if (!fundraiserId) {
+    return;
+  }
   if (req.method === "DELETE") {
-    const { fundraiserId } = req.body;
     const fundraiser = await prisma.fundraiser.delete({
       where: { id: fundraiserId },
     });
     res.json(fundraiser);
   } else if (req.method === "GET") {
-    if (!req.query.id) {
-      return;
-    }
     const fund = await prisma.fundraiser.findUnique({
       where: {
-        id: req.query.id.toString(),
+        id: fundraiserId,
       },
     });
     res.json(fund);
